+++ packages/components/accordion/src/accordion-item.tsx
@@ -1,12 +1,14 @@
 import type {Variants} from "framer-motion";
+import type {ReactNode} from "react";
+import type {UseAccordionItemProps} from "./use-accordion-item";
 
 import {forwardRef} from "@heroui/system";
-import {useMemo, ReactNode} from "react";
+import {useMemo} from "react";
 import {ChevronIcon} from "@heroui/shared-icons";
 import {AnimatePresence, LazyMotion, m, useWillChange} from "framer-motion";
 import {TRANSITION_VARIANTS} from "@heroui/framer-utils";
 
-import {UseAccordionItemProps, useAccordionItem} from "./use-accordion-item";
+import {useAccordionItem} from "./use-accordion-item";
 
 export interface AccordionItemProps extends UseAccordionItemProps {}
 
@@ -54,7 +56,11 @@ const AccordionItem = forwardRef<"button", AccordionItemProps>((props, ref) => {
 
   const content = useMemo(() => {
     if (disableAnimation) {
-      return <div {...getContentProps()}>{children}</div>;
+      if (keepContentMounted) {
+        return <div {...getContentProps()}>{children}</div>;
+      }
+
+      return isOpen && <div {...getContentProps()}>{children}</div>;
     }
 
     const transitionVariants: Variants = {