+++ packages/components/listbox/src/use-listbox-item.ts
@@ -1,18 +1,19 @@
 import type {ListboxItemBaseProps} from "./base/listbox-item-base";
 import type {MenuItemVariantProps} from "@heroui/theme";
+import type {Node} from "@react-types/shared";
+import type {ListState} from "@react-stately/list";
+import type {HTMLHeroUIProps, PropGetter} from "@heroui/system";
 
 import {useMemo, useRef, useCallback} from "react";
 import {listboxItem} from "@heroui/theme";
-import {HTMLHeroUIProps, mapPropsVariants, PropGetter, useProviderContext} from "@heroui/system";
+import {mapPropsVariants, useProviderContext} from "@heroui/system";
 import {useFocusRing} from "@react-aria/focus";
-import {Node} from "@react-types/shared";
 import {filterDOMProps} from "@heroui/react-utils";
-import {clsx, dataAttr, objectToDeps, removeEvents, warn} from "@heroui/shared-utils";
+import {clsx, dataAttr, objectToDeps, removeEvents} from "@heroui/shared-utils";
 import {useOption} from "@react-aria/listbox";
 import {mergeProps} from "@react-aria/utils";
 import {useHover, usePress} from "@react-aria/interactions";
 import {useIsMobile} from "@heroui/use-is-mobile";
-import {ListState} from "@react-stately/list";
 
 interface Props<T extends object> extends ListboxItemBaseProps<T> {
   item: Node<T>;
@@ -45,7 +46,7 @@ export function useListboxItem<T extends object>(originalProps: UseListboxItemPr
     onPressStart,
     onPressEnd,
     onPressChange,
-    onClick: deprecatedOnClick,
+    onClick,
     shouldHighlightOnFocus,
     hideSelectedIcon = false,
     isReadOnly = false,
@@ -67,16 +68,10 @@ export function useListboxItem<T extends object>(originalProps: UseListboxItemPr
 
   const isMobile = useIsMobile();
 
-  if (deprecatedOnClick && typeof deprecatedOnClick === "function") {
-    warn(
-      "onClick is deprecated, please use onPress instead. See: https://github.com/heroui-inc/heroui/issues/4292",
-      "ListboxItem",
-    );
-  }
-
   const {pressProps, isPressed} = usePress({
     ref: domRef,
     isDisabled: isDisabled,
+    onClick,
     onPress,
     onPressUp,
     onPressStart,
@@ -130,9 +125,6 @@ export function useListboxItem<T extends object>(originalProps: UseListboxItemPr
   const getItemProps: PropGetter = (props = {}) => ({
     ref: domRef,
     ...mergeProps(
-      {
-        onClick: deprecatedOnClick,
-      },
       itemProps,
       isReadOnly ? {} : mergeProps(focusProps, pressProps),
       hoverProps,